Removed all unnecessary tempo management
This encompasses every method that is like 'updateTempo' or whatever.
Only one class needs to know the current tempo, in this case it's
StateMachine.  Other classes can just read the tempo from here and
update that value only.  The double value is bound to the text in the
tempo box, so there is no need to manage that either.

Note: the field 'tempo' in StaffSequence is NOT the current tempo, it is
the tempo bound to a particular song sequence (what is saved in the
file).  I did not remove this field.