Initialize playbars in FXML
I'm removing a lot of stuff so I feel like I should explain, but I'm
working towards moving all the UI stuff to the FXML and the controller.

My understanding of MCV (or MVP or MVVM or whatever) is that the UI and
the backend don't interact directly but through the model, which in
JavaFX is implemented with a bunch of observable properties.
Interactions are set up by the controller, including user actions.

Probably the code I keep adding into the controller should be moved
elsewhere someday---for now I just don't see a better place for it.