Fix freeze issue for some situations
All GUI operations launched at the start of a song are now handled
through binders, so there is no need to wait for queue==0 at the
beginning.

I kept the check just before switching to the next song or looping,
for safety and for preserving current performances.

Also, queue is set to 0 when the player is launched. Before this there
could be some freezes resulting from an operation that failed in the
*previous* song execution.